pyproject.toml
[build-system]
requires = ["setuptools>=69.3", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-entity-management"
version = "0.1.0"
description = "A Django app for entity management with temporal data support"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name", email = "yourname@example.com" }]
license = "BSD-3-Clause"
keywords = ["django", "rest", "api", "temporal", "python"]
classifiers = [
  "Environment :: Web Environment",
  "Framework :: Django",
  "Framework :: Django :: 5.0",
  "Intended Audience :: Developers",

  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

# [project.urls]
# Homepage = "https://github.com/yourusername/django-entity-management"
# Repository = "https://github.com/yourusername/django-entity-management.git"
# Documentation = "https://yourusername.github.io/django-entity-management"
dependencies = [
  "django>=5.0",
  "djangorestframework>=3.14",
  "psycopg>=3.2",
  "python-dateutil>=2.9",
  "pytz>=2024.2",
  "django-environ>=0.11",
]

[tool.setuptools.packages.find]
include = ["app*", "entity*", "services*"]
exclude = ["typings*"]


[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-django>=4.7",
  "pytest-cov>=4.1",
  "factory-boy>=3.3",
  "freezegun>=1.5",
  "ruff>=0.6.7",
  "black>=24.8.0",
  "mypy>=1.8",
  "django-stubs>=4.2",
  "djangorestframework-stubs>=3.14",
  "types-python-dateutil>=2.8",
  "pre-commit>=3.6",
]
docs = ["sphinx>=7.0", "sphinx-rtd-theme>=1.3"]

[tool.ruff]
target-version = "py310"
line-length = 120
exclude = ["migrations", "venv", ".git", "__pycache__"]

[tool.ruff.lint]
select = [
  "E",
  "W",
  "F",
  "I",
  "B",
  "C4",
  "UP",
  "ARG",
  "SIM",
  "TCH",
  "PIE",
  "A",
  "COM",
  "C90",
  "ICN",
  "G",
  "INP",
  "T20",
  "PYI",
  "PT",
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "TID",
  "YTT",
  "ANN",
  "ASYNC",
  "EM",
  "EXE",
  "FA",
  "ISC",
  "DJ",
]
ignore = [
  "ANN101",
  "ANN102",
  "ANN401",
  "COM812",
  "ISC001",
  "Q000",
  "Q001",
  "Q002",
  "Q003",
  "TCH001",
  "TCH002",
  "TCH003",
  "UP006",
  "UP007",
  "DJ001",
  "SLF001",
]
per-file-ignores = { "__init__.py" = [
  "F401",
], "*/migrations/*" = [
  "ALL",
], "tests/*" = [
  "D",
] }

[tool.ruff.lint.isort]
known-first-party = ["entity", "services", "app"]
force-single-line = false
lines-after-imports = 2

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
plugins = ["mypy_django_plugin.main"]
# Ignore Django model field type issues
disable_error_code = ["misc"]

[tool.django-stubs]
django_settings_module = "app.settings"

[[tool.mypy.overrides]]
module = ["psycopg.*", "dateutil.*", "decouple.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["entity.models.*", "services.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["django.db.models.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["django.contrib.postgres.*"]
ignore_missing_imports = true

[tool.pyright]
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportMissingTypeStubs = false
reportAttributeAccessIssue = false
reportInvalidTypeVarUse = false
reportIncompatibleVariableOverride = false

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=app --cov-report=html"
